fix: declare explicit AllowAnonymous on public achievement endpoints (#854)
GetUserAchievementsEndpoint and GetBadgeCatalogEndpoint carried no
authorization metadata at all, making them anonymously reachable purely
by omission rather than by an explicit decision. Both serve public data,
so add .AllowAnonymous() to match the codebase's convention for public
endpoints.

Add AuthorizationConventionTests to catch this class of bug going
forward: builds a real WebApplication with every IEndpoint mapped and
asserts each resulting RouteEndpoint carries either IAllowAnonymous or
IAuthorizeData metadata. Extracted the shared WebApplication-building
helper (previously duplicated in RateLimitingConventionTests) into
EndpointTestHelper so both convention tests reuse one implementation.
