+++ backend/src/Api/Achievements/GetBadgeCatalog/v1/GetBadgeCatalogEndpoint.cs
@@ -16,6 +16,7 @@ internal sealed class GetBadgeCatalogEndpoint
 			.WithTags("Achievements")
 			.Produces<List<BadgeCatalogEntry>>()
 			.ProducesProblem(StatusCodes.Status500InternalServerError)
+			.AllowAnonymous()
 			.RequireRateLimiting(RateLimitingPolicies.Read)
 			.MapToApiVersion(1);
